RavenDB-20084: PR comments
