+++ src/Raven.Server/Documents/DocumentDatabase.cs
@@ -1571,92 +1571,6 @@ namespace Raven.Server.Documents
             }
         }
 
-        public string WhoseTaskIsIt(
-            DatabaseTopology databaseTopology,
-            IDatabaseTask configuration,
-            IDatabaseTaskStatus taskStatus,
-            bool keepTaskOnOriginalMemberNode = false)
-        {
-            return WhoseTaskIsIt(ServerStore, databaseTopology, configuration, taskStatus, NotificationCenter, keepTaskOnOriginalMemberNode);
-        }
-
-        public static string WhoseTaskIsIt(
-            ServerStore serverStore,
-            DatabaseTopology databaseTopology,
-            IDatabaseTask configuration,
-            IDatabaseTaskStatus taskStatus,
-            NotificationCenter.NotificationCenter notificationCenter,
-            bool keepTaskOnOriginalMemberNode = false)
-        {
-            var whoseTaskIsIt = databaseTopology.WhoseTaskIsIt(
-                serverStore.Engine.CurrentState, configuration,
-                getLastResponsibleNode:
-                () =>
-                {
-                    var lastResponsibleNode = taskStatus?.NodeTag;
-                    if (lastResponsibleNode == null)
-                    {
-                        // first time this task is assigned
-                        return null;
-                    }
-
-                    if (databaseTopology.AllNodes.Contains(lastResponsibleNode) == false)
-                    {
-                        // the topology doesn't include the last responsible node anymore
-                        // we'll choose a different one
-                        return null;
-                    }
-
-                    if (taskStatus is PeriodicBackupStatus)
-                    {
-                        if (databaseTopology.Rehabs.Contains(lastResponsibleNode) &&
-                            databaseTopology.PromotablesStatus.TryGetValue(lastResponsibleNode, out var status) &&
-                            (status == DatabasePromotionStatus.OutOfCpuCredits ||
-                             status == DatabasePromotionStatus.EarlyOutOfMemory ||
-                             status == DatabasePromotionStatus.HighDirtyMemory))
-                        {
-                            // avoid moving backup tasks when the machine is out of CPU credit
-                            return lastResponsibleNode;
-                        }
-                    }
-
-                    if (serverStore.LicenseManager.HasHighlyAvailableTasks() == false)
-                    {
-                        // can't redistribute, keep it on the original node
-                        RaiseAlertIfNecessary(databaseTopology, configuration, lastResponsibleNode, serverStore, notificationCenter);
-                        return lastResponsibleNode;
-                    }
-
-                    if (keepTaskOnOriginalMemberNode &&
-                        databaseTopology.Members.Contains(lastResponsibleNode))
-                    {
-                        // keep the task on the original node
-                        return lastResponsibleNode;
-                    }
-
-                    return null;
-                });
-
-            if (whoseTaskIsIt == null && taskStatus is PeriodicBackupStatus)
-                return taskStatus.NodeTag; // we don't want to stop backup process
-
-            return whoseTaskIsIt;
-        }
-
-        private static void RaiseAlertIfNecessary(DatabaseTopology databaseTopology, IDatabaseTask configuration, string lastResponsibleNode, 
-            ServerStore serverStore, NotificationCenter.NotificationCenter notificationCenter)
-        {
-            // raise alert if redistribution is necessary
-            if (notificationCenter != null &&
-                databaseTopology.Count > 1 &&
-                serverStore.NodeTag != lastResponsibleNode &&
-                databaseTopology.Members.Contains(lastResponsibleNode) == false)
-            {
-                var alert = LicenseManager.CreateHighlyAvailableTasksAlert(databaseTopology, configuration, lastResponsibleNode);
-                notificationCenter.Add(alert);
-            }
-        }
-
         public IEnumerable<DatabasePerformanceMetrics> GetAllPerformanceMetrics()
         {
             yield return TxMerger.GeneralWaitPerformanceMetrics;
